Fix ts definitions
