+++ packages/ddp/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   summary: "Meteor's latency-compensated distributed data framework",
-  version: '1.4.1'
+  version: '1.4.2-qualia.1'
 });
 
 Package.onUse(function (api) {
