+++ packages/meteor/package.js
@@ -2,7 +2,7 @@
 
 Package.describe({
   summary: "Core Meteor environment",
-  version: '1.10.2-2'
+  version: '1.10.3-qualia.1'
 });
 
 Package.registerBuildPlugin({
